udpping: increase size of control message buffer
If the control message buffer is too short to store all messages, it is
truncated. Not checking this can lead to mysterious results, e.g. false
timestamps.

The current array (256 bytes large) should not be truncated - I hope. If
however it is, there will be a log message confirming this.